requirements.txt
langchain
langgraph
langchain-core
langchain-community
langchain_groq

fastapi
uvicorn
streamlit

python-dotenv
requests
httpx
pydantic
langsmith
